NLP
This Review dataset consists of more than 500,000 rows of user reviews on amazon’s fine food
products sold everyday for the time span of 10 years. We will try to predict the reviews based on
the comments using Topic modelling and LDA (Latent Dirichlet Allocation). In this assignment
we will explore the dataset to get some interesting findings for the model to execute. We will use
Topic Modelling to gain general understanding about the what the reviews are about, and their
ratings bases on the wordcount